Add some missing sections in the parser scaffold
And move section into the main module mod.rs for simplicity.
